requirements.txt
# Python 3.13+ compatible versions
Flask>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.6
chromadb>=0.4.22
python-dotenv>=1.0.0
google-generativeai>=0.3.2
numpy>=1.26.4
lxml>=5.1.0
urllib3>=2.1.0
gunicorn>=21.2.0
setuptools>=69.0.0
wheel>=0.42.0 